src/app/core/validators/inscricao-estadual-estados/parana.validator.spec.ts
import { Parana } from './parana.validator';
import { InscricaoEstadualFactory } from './inscricao-estadual.factory';

describe('classe validacao: Inscricao Estadual Parana', () => {

  const estado = InscricaoEstadualFactory.criaInstancia(InscricaoEstadualFactory.PARANA);

  it('instancia deve estar criada', () => {
    expect(estado).toBeTruthy();
  });

  it('inscricao estadual deve estar valida', () => {
    expect(estado.validar('0127387494')).toBe(true);
    expect(estado.validar('9034205458')).toBe(true);
    expect(estado.validar('5470834821')).toBe(true);
    expect(estado.validar('3785565389')).toBe(true);
  });

  it('inscricao estadual deve estar invalida', () => {
    expect(estado.validar('123461')).toBe(false);
    expect(estado.validar('31632')).toBe(false);
    expect(estado.validar('13461123462423461')).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for nulo', () => {
    expect(estado.validar(null)).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for undefined', () => {
    expect(estado.validar(undefined)).toBe(false);
  });

});

